move: route grouping 구조로 변경(#69)
