+++ src/app/(dashboard)/dashboard/[boardId]/edit/editComponents/InviteStatus.module.scss
