+++ src/app/(dashboard)/mydashboard/page.tsx
@@ -1,4 +1,3 @@
-import CommonLayout from '@/layouts/CommonLayout';
 import { InviteProvider } from '@/contexts/inviteContext';
 import InvitedBoard from '@/components/InvitedBoard';
 import MyDashboardList from '@/components/MyDashboardList';
@@ -6,13 +5,11 @@ import styles from './MyDashboard.module.scss';
 
 export default function MyDashboard() {
   return (
-    <CommonLayout>
-      <div className={styles.container}>
-        <MyDashboardList />
-        <InviteProvider>
-          <InvitedBoard />
-        </InviteProvider>
-      </div>
-    </CommonLayout>
+    <div className={styles.container}>
+      <MyDashboardList />
+      <InviteProvider>
+        <InvitedBoard />
+      </InviteProvider>
+    </div>
   );
 }
