+++ src/app/(home)/page.module.scss
