+++ src/app/(home)/signup/page.tsx
@@ -69,7 +69,7 @@ export default function SignUpPage() {
   return (
     <div className={styles.container}>
       <Link href="/">
-        <Image className={styles.logo} src="/images/fullLogo.svg" alt="logo" width={200} height={280} />
+        <Image className={styles.logo} src="/images/fullLogo.svg" alt="logo" width={200} height={280} priority />
       </Link>
       <p className={styles.greetingText}>첫 방문을 환영합니다!</p>
       <form className={styles.form} onSubmit={handleSubmit}>
@@ -94,6 +94,7 @@ export default function SignUpPage() {
           type="password"
           placeholder="비밀번호를 입력해주세요"
           name="password"
+          autoComplete="new-password"
           onChange={handleChange}
           error={isPasswordError}
           errorMessage="비밀번호를 8자 이상 입력해주세요"
@@ -104,6 +105,7 @@ export default function SignUpPage() {
           type="password"
           placeholder="비밀번호를 한번 더 입력해주세요"
           name="confirmPassword"
+          autoComplete="new-password"
           onChange={handleChange}
           required
         />
